pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "TomaHara", email = "pyonsuke0720@gmail.com" }
]
dependencies = [
    "requests>=2.32.3",
    "icalendar>=6.1.3",
    "pytz>=2025.2",
    "python-dotenv>=1.1.0",
    "flask>=3.1.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/discord_bot"]
